move create proxy and curl examples to shared util
