+++ pkg/kf/commands/utils/utils.go
@@ -20,6 +20,9 @@ import (
 	"fmt"
 	"io"
 	"log"
+	"net"
+	"net/http"
+	"net/http/httputil"
 	"os"
 
 	"github.com/google/kf/pkg/kf/commands/config"
@@ -121,4 +124,40 @@ func parseCommandLine() (args []string, flags map[string][]string) {
 		log.Fatalf("failed to unmarshal flags: %s", err)
 	}
 	return args, flags
+}
+
+// CreateProxy creates a proxy to the specified gateway with the specified host in the request header.
+func CreateProxy(w io.Writer, host, gateway string) *httputil.ReverseProxy {
+	logger := log.New(w, fmt.Sprintf("\033[34m[%s via %s]\033[0m ", host, gateway), log.Ltime)
+
+	return &httputil.ReverseProxy{
+		Director: func(req *http.Request) {
+			req.Host = host
+			req.URL.Scheme = "http"
+			req.URL.Host = gateway
+
+			logger.Printf("%s %s\n", req.Method, req.URL.RequestURI())
+		},
+		ErrorLog: logger,
+	}
+}
+
+// PrintCurlExamples lists example HTTP requests the user can send.
+func PrintCurlExamples(w io.Writer, listener net.Listener, routeHost, gateway string, withProxy bool) {
+	if withProxy {
+		fmt.Fprintf(w, "Forwarding requests from %s to %s with host %s\n", listener.Addr(), gateway, routeHost)
+		fmt.Fprintln(w, "Example GET:")
+		fmt.Fprintf(w, "  curl %s\n", listener.Addr())
+		fmt.Fprintln(w, "Example POST:")
+		fmt.Fprintf(w, "  curl --request POST %s --data \"POST data\"\n", listener.Addr())
+		fmt.Fprintln(w, "Browser link:")
+		fmt.Fprintf(w, "  http://%s\n", listener.Addr())
+	} else {
+		fmt.Fprintf(w, "Requests can be sent to %s with host %s\n", gateway, routeHost)
+		fmt.Fprintln(w, "Example GET:")
+		fmt.Fprintf(w, "  curl -H \"Host: %s\" http://%s\n", routeHost, gateway)
+		fmt.Fprintln(w, "Example POST:")
+		fmt.Fprintf(w, "  curl --request POST -H \"Host: %s\" http://%s --data \"POST data\"\n", routeHost, gateway)
+	}
+	fmt.Fprintln(w)
 }
